+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Development
 ===========
 - (Fill this out as you fix issues and develop your features).
 - Add Mongo 4.0 to Travis
+- Fix querying and updating dynamic fields on DynamicEmbeddedDocuments
 
 Changes in 0.19.1
 =================